% Generated by roxygen2 (4.0.1): do not edit by hand
\name{save_destination_data}
\alias{save_destination_data}
\title{A function to check if the file is a csv and if it is to upload it to a sqlite database for safekeeping.}
\usage{
save_destination_data(file, ...)
}
\arguments{
\item{file}{read in a csv file.}
}
\description{
This function allows you to save the data into a sqlite database.
}
\examples{
save_destination_data()
}
\keyword{data}
\keyword{event}

